requirements.txt
# Base dependencies for OpenSEM
# Add more as needed per phase
numpy
pandas
pyyaml
mkdocs
mkdocs-material
pypdf
python-dotenv
google-generativeai
pygments
ipykernel
